refactor: landing: team section: split HunterTeam into smaller components
