+++ apps/platform-app/src/components/hunter-landing/HunterTeam/HunterTeam.module.scss
@@ -62,139 +62,3 @@
   line-height: 150%;
   margin: 0;
 }
-
-.teamMembersContainer {
-  --gap: 16px;
-  --column-width: 210px;
-  display: grid;
-  grid-template-rows: repeat(2, minmax(255px, auto));
-  grid-auto-flow: column;
-  grid-auto-columns: var(--column-width);
-  gap: var(--gap);
-  max-width: calc((var(--column-width) * 2.5) + (var(--gap) * 2));
-  padding-right: calc(var(--column-width) + var(--gap));
-  margin-bottom: 16px;
-  overflow-x: auto;
-  -ms-overflow-style: none; /* IE and Edge */
-  scrollbar-width: none; /* Firefox */
-  scroll-behavior: smooth; /* For smooth scrolling */
-  &::-webkit-scrollbar {
-    display: none; /* Chrome, Safari, and Opera */
-  }
-
-  @include mobile {
-    max-width: calc(100vw - 20px);
-    margin-right: -20px;
-    padding-right: calc(var(--column-width) / 2 + var(--gap));
-  }
-  @include desktop {
-    padding-right: calc(var(--column-width) / 2 + var(--gap));
-    max-width: calc((var(--column-width) * 3.5) + (var(--gap) * 3));
-  }
-}
-
-.memberCard {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  gap: 12px;
-  background-color: var(--hunter-element-bg-color);
-  border-radius: 4px;
-  padding: 20px;
-  position: relative;
-  height: 100%;
-}
-
-.avatarContainer {
-  position: relative;
-  background-color: --hunter-landing-text-color3;
-}
-
-.avatar {
-  width: 120px;
-  height: 120px;
-  border-radius: 50%;
-  object-fit: cover;
-}
-
-.statusIndicator {
-  position: absolute;
-  top: 7px;
-  right: 5px;
-}
-
-.statusDot {
-  width: 18px;
-  height: 18px;
-  background-color: #00a700;
-  border-radius: 50%;
-}
-
-.memberInfo {
-  width: 100%;
-  text-align: center;
-}
-
-.memberName {
-  color: var(--hunter-landing-text-color2);
-  font-size: 16px;
-  font-style: normal;
-  font-weight: 500;
-  line-height: 20px;
-  margin-bottom: 2px;
-}
-
-.memberPosition {
-  color: #5e28f6;
-  font-size: 14px;
-  font-style: normal;
-  font-weight: 400;
-  line-height: 20px;
-  margin-bottom: 3px;
-}
-
-.memberLocation {
-  color: #69717b;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 450;
-  line-height: 16px;
-}
-
-.availabilityInfo {
-  color: #00a700;
-  text-align: center;
-  font-size: 12px;
-  font-style: normal;
-  font-weight: 450;
-  line-height: 16px;
-}
-
-.navigation {
-  display: flex;
-  gap: 30px;
-  margin-top: 16px;
-  align-self: flex-start;
-}
-
-.navButton {
-  width: 32px;
-  height: 32px;
-  background-color: transparent;
-  color: var(--color);
-  border: none;
-  cursor: pointer;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  transition: all 0.3s ease;
-
-  &:disabled {
-    opacity: 0.5;
-    cursor: not-allowed;
-  }
-
-  &:not(:disabled):hover {
-    background-color: var(--hunter-landing-bg-color3);
-  }
-}
