+++ apps/platform-app/src/components/hunter-landing/HunterTeam/TeamMemberCard.module.scss
@@ -1,0 +1,76 @@
+.memberCard {
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  gap: 12px;
+  background-color: var(--hunter-element-bg-color);
+  border-radius: 4px;
+  padding: 20px;
+  position: relative;
+  height: 100%;
+}
+
+.avatarContainer {
+  position: relative;
+  background-color: --hunter-landing-text-color3;
+}
+
+.avatar {
+  width: 120px;
+  height: 120px;
+  border-radius: 50%;
+  object-fit: cover;
+}
+
+.statusIndicator {
+  position: absolute;
+  top: 7px;
+  right: 5px;
+}
+
+.statusDot {
+  width: 18px;
+  height: 18px;
+  background-color: #00a700;
+  border-radius: 50%;
+}
+
+.memberInfo {
+  width: 100%;
+  text-align: center;
+}
+
+.memberName {
+  color: var(--hunter-landing-text-color2);
+  font-size: 16px;
+  font-style: normal;
+  font-weight: 500;
+  line-height: 20px;
+  margin-bottom: 2px;
+}
+
+.memberPosition {
+  color: #5e28f6;
+  font-size: 14px;
+  font-style: normal;
+  font-weight: 400;
+  line-height: 20px;
+  margin-bottom: 3px;
+}
+
+.memberLocation {
+  color: #69717b;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 450;
+  line-height: 16px;
+}
+
+.availabilityInfo {
+  color: #00a700;
+  text-align: center;
+  font-size: 12px;
+  font-style: normal;
+  font-weight: 450;
+  line-height: 16px;
+}
